--- a/NEWS
+++ b/NEWS
@@ -1,5 +1,15 @@
+Broker 1.5.0
+============
+
+TODO: No important changes yet
+
 Broker 1.4.0
 ============
+
+- Adds a new ``broker::store_event`` API that can be used to observe
+  data store modifications.
+
+- Adds support for Windows platform.
 
 - RocksDB support is now opt-in instead of automatically detected and used
   at configuration-time.  Use the ``--enable-rocksdb`` and