--- a/bindings/python/_broker.cpp
+++ b/bindings/python/_broker.cpp
@@ -124,7 +124,8 @@ PYBIND11_MODULE(_broker, m) {
     .def("drop_all_on_destruction", &broker::publisher::drop_all_on_destruction)
     .def("publish", (void (broker::publisher::*)(broker::data d)) &broker::publisher::publish)
     .def("publish_batch",
-       [](broker::publisher& p, std::vector<broker::data> xs) { p.publish(xs); });
+       [](broker::publisher& p, std::vector<broker::data> xs) { p.publish(xs); })
+    .def("reset", &broker::publisher::reset);
 
   using subscriber_base = broker::subscriber_base<broker::subscriber::value_type>;
   using topic_data_pair = std::pair<broker::topic, broker::data>;
@@ -190,7 +191,8 @@ PYBIND11_MODULE(_broker, m) {
 
   py::class_<broker::subscriber, subscriber_base>(m, "Subscriber")
     .def("add_topic", &broker::subscriber::add_topic)
-    .def("remove_topic", &broker::subscriber::remove_topic);
+    .def("remove_topic", &broker::subscriber::remove_topic)
+    .def("reset", &broker::subscriber::reset);
 
   py::bind_vector<std::vector<broker::status_subscriber::value_type>>(m, "VectorStatusSubscriberValueType");
 
@@ -222,7 +224,8 @@ PYBIND11_MODULE(_broker, m) {
          [](broker::status_subscriber& ep) -> std::vector<broker::status_subscriber::value_type> {
 	   return ep.poll(); })
     .def("available", &broker::status_subscriber::available)
-    .def("fd", &broker::status_subscriber::fd);
+    .def("fd", &broker::status_subscriber::fd)
+    .def("reset", &broker::status_subscriber::reset);
 
   py::class_<broker::status_subscriber::value_type>(status_subscriber, "ValueType")
     .def("is_error",