--- a/bindings/python/store.cpp
+++ b/bindings/python/store.cpp
@@ -48,8 +48,8 @@ void init_store(py::module& m) {
     .def("increment", &broker::store::increment)
     .def("decrement", &broker::store::decrement)
     .def("append", &broker::store::append)
-    .def("insert_into", (void (broker::store::*)(broker::data, broker::data, broker::optional<broker::timespan>) const) &broker::store::insert_into)
-    .def("insert_into", (void (broker::store::*)(broker::data, broker::data, broker::data, broker::optional<broker::timespan>) const) &broker::store::insert_into)
+    .def("insert_into", (void (broker::store::*)(broker::data, broker::data, broker::optional<broker::timespan>)) &broker::store::insert_into)
+    .def("insert_into", (void (broker::store::*)(broker::data, broker::data, broker::data, broker::optional<broker::timespan>)) &broker::store::insert_into)
     .def("remove_from", &broker::store::remove_from)
     .def("push", &broker::store::push)
     .def("pop", &broker::store::pop);