--- a/ci/fedora-32/Dockerfile
+++ b/ci/fedora-32/Dockerfile
@@ -1,4 +1,4 @@
-FROM fedora:30
+FROM fedora:32
 
 RUN yum -y install \
     git \