--- a/configure
+++ b/configure
@@ -123,7 +123,7 @@ while [ $# -ne 0 ]; do
             append_cache_entry CMAKE_BUILD_TYPE     STRING $optarg
             ;;
         --enable-debug)
-            append_cache_entry ENABLE_DEBUG         BOOL   true
+            append_cache_entry BROKER_ENABLE_DEBUG  BOOL   true
             ;;
         --enable-static)
             append_cache_entry ENABLE_STATIC        BOOL   true