--- a/configure
+++ b/configure
@@ -123,7 +123,7 @@ while [ $# -ne 0 ]; do
             append_cache_entry CMAKE_BUILD_TYPE     STRING $optarg
             ;;
         --enable-debug)
-            append_cache_entry BROKER_ENABLE_DEBUG  BOOL   true
+            append_cache_entry CAF_ENABLE_RUNTIME_CHECKS BOOL true
             ;;
         --enable-static)
             append_cache_entry ENABLE_STATIC        BOOL   true
@@ -147,7 +147,7 @@ while [ $# -ne 0 ]; do
             append_cache_entry BROKER_DISABLE_TESTS BOOL    true
             ;;
         --with-caf=*)
-            append_cache_entry CAF_ROOT_DIR         PATH    $optarg
+            append_cache_entry CAF_ROOT             PATH    $optarg
             ;;
         --with-openssl=*)
             append_cache_entry OPENSSL_ROOT_DIR     PATH    $optarg