--- a/include/broker/address.hh
+++ b/include/broker/address.hh
@@ -9,6 +9,8 @@
 #include <caf/ip_address.hpp>
 #include <caf/string_view.hpp>
 
+#include "broker/detail/is_legacy_inspector.hh"
+
 namespace broker {
 
 /// Stores an IPv4 or IPv6 address.
@@ -71,10 +73,7 @@ public:
     return addr_.compare(other.addr_);
   }
 
-  template <class Inspector>
-  friend typename Inspector::result_type inspect(Inspector& f, address& x) {
-    return f(x.addr_);
-  }
+  // -- conversion support -----------------------------------------------------
 
   friend bool convert(const address& a, std::string& str) {
     str = to_string(a.addr_);
@@ -87,15 +86,27 @@ public:
     return true;
   }
 
+  // -- inspection support -----------------------------------------------------
+
+  // We transparently expose the member variable. Hence, broker::address and
+  // caf::ip_address are one and the same to CAF inspectors.
+
+  template <class Inspector>
+  friend typename Inspector::result_type inspect(Inspector& f, address& x) {
+    if constexpr (detail::is_legacy_inspector<Inspector>)
+      return f(x.addr_);
+    else
+      return f.apply_object(x);
+  }
+
+  template <class Inspector>
+  friend bool inspect_value(Inspector& f, address& x) {
+    return f.apply_value(x.addr_);
+  }
+
 private:
   caf::ip_address addr_;
 };
-
-/// @relates address
-bool convert(const std::string& str, address& a);
-
-/// @relates address
-bool convert(const address& a, std::string& str);
 
 } // namespace broker