This is a followup to #89 and gets rebased once that hits `master` before this PR can leave *draft* status.

- [x] Require CAF 0.18 via CMake.
- [x] Update submodule version of CAF and migrate to the new CMake options / targets.
- [ ] Remove all `ifdef` statements for supporting CAF < 0.18.