@jsiwek 2020-12-10
@Neverlord since CAF 0.18 is merged already from #153, can this get closed as obsoleted or plan to consolidate with #89 ?  Probably would go with #89 as canonical since it contains all more context/discussion.